requirements.txt
Flask
pdfminer.six
openai
python-dotenv
Werkzeug
PyPDF2
PyMuPDF 